feat: rename CopyButton to Copy
